requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
opencv-python==4.8.1.78
numpy==1.26.2
tensorflow==2.15.0
scikit-learn==1.3.2
Pillow==10.1.0
bcrypt==4.1.2
python-dotenv==1.0.0
werkzeug==3.0.1
matplotlib==3.8.2
scipy==1.11.4
